Add files via upload
This is an adaptation of Kruskal's algorithm but used for a Maximum spanning tree. 

Problem statement: Gary needs sports equipment for his newly opened shop in city which he can acquire from N number of shops. Every shop is accessible from every other shop. There is a Path P from between shops and there is a cost to travel on the path from one shop to another. But the cost is charged only the first time. For instance, if Cost of traveling from shop 1 to 2 is 20, and Gary travels from shop 1 to 2 and back to 1, he will be charged only 20 for the first time. We need to find the maximum cost path for every shop without forming a cycle.